reorganized types into safeguards_types.go
